# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SFMA Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found

    patch:
      operationId: updateMember
      tags: [members]
      summary: Update member (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/ban:
    post:
      operationId: banMember
      tags: [members]
      summary: Ban a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Banned
  /members/{id}/avatar:
    post:
      operationId: uploadAvatar
      tags: [members]
      summary: Upload member avatar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarInput"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /messages/{id}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      summary: Add/toggle emoji reaction to a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /polls:
    get:
      operationId: listPolls
      tags: [polls]
      summary: List all polls
      responses:
        "200":
          description: Polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"

    post:
      operationId: createPoll
      tags: [polls]
      summary: Create a poll (Lords only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PollInput"
      responses:
        "201":
          description: Poll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
  /polls/{id}/vote:
    post:
      operationId: votePoll
      tags: [polls]
      summary: Vote on a poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, username, grade, memberId, joinedAt, reputation]
      properties:
        id:
          type: string
        username:
          type: string
        grade:
          type: string
          enum: [Fondateur_Suprême, Co-Fondateur, Lord, Administrateur, Recruteur, Modérateur, Membre_Élite, Membre, Nouveau_Membre]
        gradeEmoji:
          type: string
        memberId:
          type: string
        country:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        reputation:
          type: integer
        joinedAt:
          type: string
        isBanned:
          type: boolean
        badges:
          type: array
          items:
            type: string
    MemberUpdate:
      type: object
      properties:
        grade:
          type: string
        bio:
          type: string
        country:
          type: string
    AvatarInput:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
...
    Message:
      type: object
      required: [id, groupId, content, author, createdAt]
      properties:
        id:
          type: string
        groupId:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/Member"
        replyTo:
          type: ["object", "null"]
          properties:
            id:
              type: string
            content:
              type: string
            authorUsername:
              type: string
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        createdAt:
          type: string
        mentions:
          type: array
          items:
            type: string
...
    Reaction:
      type: object
      required: [emoji, count, userIds]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userIds:
          type: array
          items:
            type: string
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
...
    Poll:
      type: object
      required: [id, question, options, createdAt, createdBy]
      properties:
        id:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        createdAt:
          type: string
        createdBy:
          type: string
        closedAt:
          type: ["string", "null"]
        totalVotes:
          type: integer
        userVotedOption:
          type: ["string", "null"]
    PollOption:
      type: object
      required: [id, text, votes]
      properties:
        id:
          type: string
        text:
          type: string
        votes:
          type: integer
        percentage:
          type: number
    PollInput:
      type: object
      required: [question, options]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
    VoteInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string
...