# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SFMA Community Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  grade?: string;
  bio?: string;
  country?: string;
}