# lib/api-zod/src/generated/types/poll.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SFMA Community Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PollOption } from './pollOption';

export interface Poll {
  id: string;
  question: string;
  options: PollOption[];
  createdAt: string;
  createdBy: string;
  /** @nullable */
  closedAt?: string | null;
  totalVotes?: number;
  /** @nullable */
  userVotedOption?: string | null;
}